Stop the metronome when changing its settings
This is the best option for now - since the metronome works by pre-scheduling
one pass through all bars, it can't be changed on the fly, and re-starting it
everytime the user clicks something would be annoying. Instead, we'll stop on
the first click, then allow the user to make their changes and restart the
metronome - under the hypothesis that most users will make more than one change
at a time.